pyproject.toml
[project]
name = "nss-safedata-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "diffprivlib==0.6.1",
    "fastparquet==2024.5.0",
    "numpy==1.26.4",
    "pandas==2.2.2",
    "pyarrow==16.1.0",
    "pydantic>=2.12.3",
    "pytest==8.2.2",
    "pytest-mock>=3.15.1",
    "pyyaml==6.0.1",
    "reportlab>=4.4.4",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "sdv==1.1.0",
]
